add probably more correct spawnDualSnowSprayEffect definition

+++ include/race/spray_particles.h
@@ -10,6 +10,6 @@ void spawnImpactStar(Vec3i *arg0);
 void spawnCharacterAttackEffect(Player *player);
 void spawnCharacterAttackEffectByType(Player *player, s32 characterId);
 void spawnGlintEffect(void *arg0);
-void spawnDualSnowSprayEffect(Vec3i *pos1, Vec3i *pos2, Vec3i *velocity, s16 slotIndex, s32 characterId);
+void spawnDualSnowSprayEffect(Vec3i *pos1, Vec3i *pos2, Vec3i *velocity, s32 slotIndex, s32 characterId);
 void spawnDualSnowSprayEffect_SingleSlot(Vec3i *pos1, Vec3i *pos2, Vec3i *velocity, s32 particleType);
 void spawnSprayEffect(Vec3i *arg0, Vec3i *arg1, s32 arg2);
